implement phase factors into Laplacian, phase factors of the non-local potential still needed

+++ src/debug_output.hxx
@@ -9,11 +9,11 @@
 template <typename real_t>
 int dump_to_file(
       char const *filename
-    , int const N // number of rows
+    , size_t const N // number of rows
     , real_t const y_data[] // 2D-data array [N*Stride]
     , double *x_axis=nullptr // x-axis[N], replaced by 0,1,2,... if 0x0
-    , int const Stride=1 // Stride, default=1 for 1D functions
-    , int const M=1 // number of columns to display
+    , size_t const Stride=1 // Stride, default=1 for 1D functions
+    , size_t const M=1 // number of columns to display
     , char const *title=nullptr // title line in the file
     , int const echo=0 // report writing to stdout, 0: suppress report
 ) {
